git: mailsplit
